requirements.txt
attrdict==2.0.1
einops==0.8.1
flash_attn==2.8.3
gradio==6.0.0
huggingface_hub==0.36.0
json_repair==0.54.1
langchain_core==1.1.0
Markdown==3.6
matplotlib==3.10.7
mdtex2html==1.3.0
mplfinance==0.12.10b0
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
pydantic==2.12.4
Pygments==2.19.2
timm==1.0.22
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.64.1
transformers==4.38.2
xformers==0.0.33.post1
yfinance==0.2.66
